Added destructor debug messages. Add macro to enable/disable automatic help generation. Removed static global variables in Rx class and put them in Rx.Run() instead. This fixed the issues with some of the unit tests failing.

+++ src/Clide-Tx.cpp
@@ -312,7 +312,7 @@ namespace Clide
 		}
 		
 		// Copy parameters into cmd string
-		for(x = 0; x < foundCmd->numParams; x++)
+		for(x = 0; (uint32_t)x < foundCmd->numParams; x++)
 		{
 			strcpy(foundCmd->paramA[x]->value, _argsPtr[optind + x]);
 		}
